Create manual_img_digitizer
This script uses the opencv library to perform a scan of a rectangular image. When run, it will open an image file, display it in a resizable window, and allow the user to select the four points that refer to the borders of the image. It then calculates a homographic matrix based on the points clicked, and uses this matrix to transform the input image into a new image that is rectified and cropped to the bounded region. Finally, it displays both the transformed image and an image with previews of the selected edges and the computed homography.